moved c prototype parser to extra file for testing
